requirements.txt
Flask
google-cloud-bigquery
google-cloud-logging
google-api-python-client
